Css for Explore Page and Profile Page DONE. Also, fixed the incorrect route to components ExploreFoodScreen and ExploreDrinkScreen.
