+++ src/App.js
@@ -5,8 +5,8 @@ import './App.css';
 import Login from './pages/Login/Login';
 import MainFoodsScreen from './pages/MainFoodsScreen';
 import MainDrinksScreen from './pages/MainDrinksScreen';
-import Explore from './pages/Explore';
-import Profile from './pages/Profile';
+import Explore from './pages/Explore/Explore';
+import Profile from './pages/Profile/Profile';
 import FavoriteRecipes from './pages/FavoriteRecipes';
 import NotFoundPage from './pages/NotFoundPage';
 import RecipesMade from './pages/RecipesMade';
@@ -28,8 +28,8 @@ function App() {
             <Route exact path="/comidas/:id/in-progress" component={null} />
             <Route exact path="/bebidas/:id/in-progress" component={null} />
             <Route exact path="/explorar" component={Explore} />
-            <Route exact path="/explorar/comidas" component={ExploreDrinksScreen} />
-            <Route exact path="/explorar/bebidas" component={ExploreFoodsScreen} />
+            <Route exact path="/explorar/comidas" component={ExploreFoodsScreen} />
+            <Route exact path="/explorar/bebidas" component={ExploreDrinksScreen} />
             <Route exact path="/explorar/comidas/ingredientes" component={null} />
             <Route exact path="/explorar/bebidas/ingredientes" component={null} />
             <Route exact path="/explorar/comidas/area" component={null} />
