+++ src/components/Details/ingredientsToshow.js
@@ -7,7 +7,7 @@ const ingredientsToshow = (el) => {
       obj.push(
         <div data-testid={`${index}-ingredient-name-and-measure`}>
           { el[`strIngredient${index}`]}
-          :
+          -
           {el[`strMeasure${index}`] }
         </div>,
       );
