+++ src/components/Header/Header.style.css
@@ -7,19 +7,22 @@
 }
 
 .title {
-  align-items: center;
-  color: black;
-  display: flex;
+  left: 80px;
+  position: absolute;
+  top: 10px;
+  width: 200px;
+}
+
+.title p{
+  color: #333;
   font-family: Yellowtail, cursive;
-  font-size: 40px;
+  font-size: 30px;
   font-style: normal;
   font-weight: 100;
-  height: 35px;
-  left: 123px;
-  line-height: 35px;
-  position: absolute;
-  top: 12px;
-  width: 119px;
+  margin: 0;
+  padding: 0;
+  text-align: center;
+  width: 200px;
 }
 
 .profile-icon {
