--- src/pages/Profile.js
@@ -1,33 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
-import Header from '../components/Header/Header';
-import Footer from '../components/Footer/Footer';
-
-export default function Profile() {
-  const [user, setUser] = useState('');
-
-  useEffect(() => {
-    setUser(JSON.parse(localStorage.getItem('user')));
-  }, []);
-
-  return (
-    <div>
-      {Header('Perfil', false)}
-      <div className="holder">
-        <h2 data-testid="profile-email">{user.email}</h2>
-        <Link to="/receitas-favoritas">
-          <button data-testid="profile-favorite-btn">Receitas Favoritas</button>
-        </Link>
-        <Link to="/receitas-feitas">
-          <button data-testid="profile-done-btn">Receitas Feitas</button>
-        </Link>
-        <Link to="/">
-          <button onClick={() => localStorage.clear()} data-testid="profile-logout-btn">
-            Sair
-          </button>
-        </Link>
-      </div>
-      <Footer />
-    </div>
-  );
-}
